requirements.txt
pandas==1.5.2
numpy==1.21.5
scikit-learn==1.0.2
xgboost
joblib==1.3.2
imbalanced-learn
flask
PyYAML==6.0.1
google-genai
python-dotenv